popularity-tracker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
names-dataset==3.1.0
numpy==1.26.4
openai==1.23.6
pandas==2.2.2
pycountry==23.12.11
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
